Update R_make.m
Made it so that freq and S_R have lengths that correspond with N (before, they corresponded to a doubly long time series). This was to more easily facilitate an important change in red_tide.m that reduces the double-counting of noise variance in both noise and signal components of the model.